[QR-045] Implement Clean Architecture (Revision) (#75)
* Change the code structure
