--- app/src/main/java/com/bangkit/bisamerchant/core/data/Result.kt
@@ -1,7 +1,0 @@
-package com.bangkit.bisamerchant.core.data
-
-sealed class Result<out R> private constructor() {
-    data class Success<out T>(val data: T) : Result<T>()
-    data class Error(val error: String) : Result<Nothing>()
-    object Loading : Result<Nothing>()
-}
