--- app/src/main/java/com/bangkit/bisamerchant/core/data/model/Payment.kt
@@ -1,23 +1,0 @@
-package com.bangkit.bisamerchant.core.data.model
-
-import com.google.gson.annotations.SerializedName
-
-data class Payment(
-    @field:SerializedName("amount")
-    val amount: Long,
-
-    @field:SerializedName("id")
-    val id: Long? = null,
-
-    @field:SerializedName("merchantId")
-    val merchantId: String? = null,
-
-    @field:SerializedName("payerId")
-    val payerId: String,
-
-    @field:SerializedName("timestamp")
-    val timestamp: Long? = null,
-
-    @field:SerializedName("trxType")
-    val trxType: String = "PAYMENT"
-)
