--- app/src/main/java/com/bangkit/bisamerchant/core/domain/model/DetailTransaction.kt
@@ -1,12 +1,0 @@
-package com.bangkit.bisamerchant.core.domain.model
-
-data class DetailTransaction(
-	val amount: Long? = null,
-	val merchantId: String? = null,
-	val payerId: String? = null,
-	val bankAccountNo: Long? = null,
-	val bankInst: String? = null,
-	val trxType: String? = null,
-	val id: String? = null,
-	val timestamp: Long? = null
-)
