--- app/src/main/java/com/bangkit/bisamerchant/core/domain/model/MessageNotif.kt
@@ -1,7 +1,0 @@
-package com.bangkit.bisamerchant.core.domain.model
-
-data class MessageNotif(
-	val title: String,
-	val body: String,
-	val subText: String,
-)
