--- app/src/main/java/com/bangkit/bisamerchant/core/domain/model/Transaction.kt
@@ -1,8 +1,0 @@
-package com.bangkit.bisamerchant.core.domain.model
-
-data class Transaction(
-	val amount: Long,
-	val trxType: String? = null,
-	val id: String? = null,
-	val timestamp: Long? = null
-)
