--- app/src/main/java/com/bangkit/bisamerchant/core/domain/usecase/MerchantUseCase.kt
@@ -1,19 +1,0 @@
-package com.bangkit.bisamerchant.core.domain.usecase
-
-import com.bangkit.bisamerchant.core.domain.model.Merchant
-import com.google.firebase.firestore.ListenerRegistration
-import com.google.firebase.firestore.QuerySnapshot
-
-interface MerchantUseCase {
-    suspend fun observeMerchantActive(callback: (Merchant) -> Unit): ListenerRegistration
-    suspend fun observeMerchants(callback: (List<Merchant>) -> Unit): ListenerRegistration
-    fun stopObserving()
-    fun processMerchantActiveQuerySnapshot(querySnapshot: QuerySnapshot): Merchant
-    fun processMerchantsQuerySnapshot(querySnapshot: QuerySnapshot): List<Merchant>
-    suspend fun changeMerchantStatus(id: String?)
-    suspend fun getMerchantId(): String
-    suspend fun saveMerchantId(id: String)
-    suspend fun saveAmountHide(hide: Boolean)
-    suspend fun getAmountHide(): Boolean
-    suspend fun deleteMerchant()
-}
