+++ app/src/main/java/com/bangkit/bisamerchant/domain/home/model/MessageNotif.kt
@@ -1,0 +1,7 @@
+package com.bangkit.bisamerchant.domain.home.model
+
+data class MessageNotif(
+	val title: String,
+	val body: String,
+	val subText: String,
+)
