+++ app/src/main/java/com/bangkit/bisamerchant/domain/home/model/Payment.kt
@@ -1,0 +1,10 @@
+package com.bangkit.bisamerchant.domain.home.model
+
+data class Payment(
+    val amount: Long,
+    val id: Long? = null,
+    val merchantId: String? = null,
+    val payerId: String,
+    val timestamp: Long? = null,
+    val trxType: String = "PAYMENT"
+)
