+++ app/src/main/java/com/bangkit/bisamerchant/domain/home/model/Transaction.kt
@@ -1,0 +1,8 @@
+package com.bangkit.bisamerchant.domain.home.model
+
+data class Transaction(
+	val amount: Long,
+	val trxType: String? = null,
+	val id: String? = null,
+	val timestamp: Long? = null
+)
