+++ app/src/main/java/com/bangkit/bisamerchant/domain/merchantsetting/repository/IMerchantSettingRepository.kt
@@ -1,0 +1,8 @@
+package com.bangkit.bisamerchant.domain.merchantsetting.repository
+
+import com.bangkit.bisamerchant.domain.merchantsetting.model.Merchant
+import com.google.firebase.firestore.ListenerRegistration
+
+interface IMerchantSettingRepository {
+    suspend fun getMerchantActive(callback: (Merchant) -> Unit): ListenerRegistration
+}
