+++ app/src/main/java/com/bangkit/bisamerchant/presentation/merchantsetting/viewmodel/MerchantSettingViewModel.kt
@@ -1,0 +1,31 @@
+package com.bangkit.bisamerchant.presentation.merchantsetting.viewmodel
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.bangkit.bisamerchant.domain.merchantsetting.model.Merchant
+import com.bangkit.bisamerchant.domain.merchantsetting.usecase.GetMerchantActive
+import com.google.firebase.firestore.ListenerRegistration
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.launch
+import javax.inject.Inject
+
+@HiltViewModel
+class MerchantSettingViewModel @Inject constructor(
+    private val getMerchantActive: GetMerchantActive
+) : ViewModel() {
+
+    private val _merchant = MutableLiveData<Merchant>()
+    val merchant: LiveData<Merchant> get() = _merchant
+
+    private var listenerRegistration: ListenerRegistration? = null
+
+    fun getMerchantActive() {
+        viewModelScope.launch {
+            listenerRegistration = getMerchantActive.execute { merchant ->
+                _merchant.value = merchant
+            }
+        }
+    }
+}
